+++ src/components/FilterBox/FilterBox.css
@@ -1,0 +1,12 @@
+select {
+  height: 44px;
+  min-width: 125px;
+
+  padding: 8px;
+
+  border: 1px solid #d0d5dd;
+  border-radius: 8px;
+  background: transparent;
+
+  font-size: 16px;
+}
