+++ src/components/LunchPickerTab/LunchPickerTab.css
@@ -1,0 +1,30 @@
+.lunch-picker-tab {
+  display: flex;
+
+  padding: 0 16px;
+  margin: 24px 0;
+}
+
+.tab-item {
+  width: 50%;
+
+  padding: 9px 0;
+
+  border: none;
+  background: var(--grey-100);
+  color: var(--grey-300);
+  border-bottom: 2px solid var(--grey-200);
+
+  text-align: center;
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 24px;
+  letter-spacing: 0.5px;
+
+  cursor: pointer;
+}
+
+.tab-item--checked {
+  color: var(--primary-color);
+  border-bottom: 2px solid var(--primary-color);
+}
