+++ src/components/modal/RestaurantDetailModal/RestaurantDetailModal.css
@@ -1,0 +1,9 @@
+.restaurant-detail-container {
+  position: relative;
+
+  display: flex;
+  flex-direction: column;
+  row-gap: 16px;
+
+  margin-bottom: 32px;
+}
