--- src/styles/style.css
@@ -1,374 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-ul,
-li {
-  list-style: none;
-}
-
-html,
-body {
-  font-family: sans-serif;
-  font-size: 16px;
-}
-
-/* Colors *****************************************/
-:root {
-  --primary-color: #ec4a0a;
-  --lighten-color: #f6a88a;
-  --grey-100: #ffffff;
-  --grey-200: #d0d5dd;
-  --grey-300: #667085;
-  --grey-400: #344054;
-  --grey-500: #000000;
-}
-
-/* Typography *************************************/
-.text-title {
-  font-size: 20px;
-  line-height: 24px;
-  font-weight: 600;
-}
-
-.text-subtitle {
-  font-size: 18px;
-  line-height: 28px;
-  font-weight: 600;
-}
-
-.text-body {
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 400;
-}
-
-.text-caption {
-  font-size: 14px;
-  line-height: 20px;
-  font-weight: 400;
-}
-
-/* GNB ********************************************/
-.gnb {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 64px;
-
-  padding: 0 16px;
-
-  background-color: var(--primary-color);
-}
-
-.gnb__title {
-  color: #fcfcfd;
-}
-
-.gnb__button {
-  height: 40px;
-
-  border: none;
-  border-radius: 8px;
-  background: transparent;
-
-  font-size: 24px;
-  cursor: pointer;
-}
-
-.gnb__button img {
-  display: block;
-  width: 40px;
-  height: 40px;
-  object-fit: contain;
-}
-
-/* 음식점 목록 *****************************************/
-
-/* 카테고리/정렬 필터 */
-.restaurant-filter-container {
-  display: flex;
-  justify-content: space-between;
-
-  padding: 0 16px;
-  margin-top: 24px;
-}
-
-.restaurant-filter-container select {
-  height: 44px;
-  min-width: 125px;
-
-  padding: 8px;
-
-  border: 1px solid #d0d5dd;
-  border-radius: 8px;
-  background: transparent;
-
-  font-size: 16px;
-}
-
-/* 음식점 목록 */
-.restaurant-list-container {
-  display: flex;
-  flex-direction: column;
-
-  padding: 0 16px;
-  margin: 16px 0;
-}
-
-.restaurant-list__default {
-  margin: 64px 0;
-
-  color: var(--grey-300);
-  text-align: center;
-}
-
-.restaurant {
-  position: relative;
-
-  display: flex;
-  align-items: flex-start;
-
-  padding: 16px 8px;
-
-  border-bottom: 1px solid #e9eaed;
-}
-
-.restaurant__category {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 64px;
-  height: 64px;
-  min-width: 64px;
-  min-height: 64px;
-
-  margin-right: 16px;
-
-  border-radius: 50%;
-  background: var(--lighten-color);
-}
-
-.category-icon {
-  width: 36px;
-  height: 36px;
-}
-
-.restaurant__info {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-
-  width: 100%;
-}
-
-.restaurant__name {
-  margin: 0;
-}
-
-.restaurant__distance {
-  color: var(--primary-color);
-}
-
-.restaurant__description {
-  display: -webkit-box;
-
-  padding-top: 8px;
-
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-}
-
-.restaurant-detail-container {
-  position: relative;
-
-  display: flex;
-  flex-direction: column;
-  row-gap: 16px;
-
-  margin-bottom: 32px;
-}
-
-.favorite__button {
-  position: absolute;
-  right: 0;
-
-  height: 32px;
-
-  border: none;
-  border-radius: 8px;
-  background: transparent;
-
-  cursor: pointer;
-}
-
-/* 음식점 추가 모달 *****************************************/
-.modal {
-  display: none;
-}
-
-.modal--open {
-  display: block;
-}
-
-.modal-backdrop {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: 1;
-
-  background: rgba(0, 0, 0, 0.35);
-}
-
-.modal-container {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  z-index: 2;
-
-  width: 100%;
-  max-height: 90vh;
-  overflow: scroll;
-
-  padding: 32px 16px;
-
-  border-radius: 8px 8px 0px 0px;
-  background: var(--grey-100);
-}
-
-.modal-title {
-  margin-bottom: 36px;
-}
-
-.form-item {
-  display: flex;
-  flex-direction: column;
-
-  margin-bottom: 36px;
-}
-
-.form-item label {
-  color: var(--grey-400);
-  font-size: 14px;
-}
-
-.form-item--required label::after {
-  padding-left: 4px;
-
-  color: var(--primary-color);
-  content: '*';
-}
-
-.form-item .help-text {
-  color: var(--grey-300);
-}
-
-.form-item input,
-.form-item textarea,
-.form-item select {
-  width: 100%;
-  padding: 8px;
-  margin: 6px 0;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  font-size: 16px;
-}
-
-.form-item textarea {
-  resize: none;
-}
-
-.form-item select {
-  height: 44px;
-
-  padding: 8px;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  color: var(--grey-300);
-}
-
-.form-item p {
-  font-size: 12px;
-  color: red;
-}
-
-input[name='name'],
-input[name='link'] {
-  height: 44px;
-}
-
-.button-container {
-  display: flex;
-}
-
-.button {
-  width: 100%;
-  height: 44px;
-
-  margin-right: 16px;
-
-  border: none;
-  border-radius: 8px;
-
-  font-weight: 600;
-  cursor: pointer;
-}
-
-.button:last-child {
-  margin-right: 0;
-}
-
-.button--secondary {
-  border: 1px solid var(--grey-300);
-  background: transparent;
-
-  color: var(--grey-300);
-}
-
-.button--primary {
-  background: var(--primary-color);
-
-  color: var(--grey-100);
-}
-
-/* 음식점 목록 탭 */
-.lunch-picker-tab {
-  display: flex;
-
-  padding: 0 16px;
-  margin: 24px 0;
-}
-
-.tab-item {
-  width: 50%;
-
-  padding: 9px 0;
-
-  border: none;
-  background: var(--grey-100);
-  color: var(--grey-300);
-  border-bottom: 2px solid var(--grey-200);
-
-  text-align: center;
-  font-size: 16px;
-  font-weight: 700;
-  line-height: 24px;
-  letter-spacing: 0.5px;
-
-  cursor: pointer;
-}
-
-.tab-item--checked {
-  color: var(--primary-color);
-  border-bottom: 2px solid var(--primary-color);
-}
